move `repr_key` from `config` to `handler`
